arch-riscv: Fix fflags behavior of float inst. in O3 CPU
According to the RISC-V spec [1]. Any float-point instructions accumulate FFLAGS register rather than write it to reflect the CSR behavior.

In the previous implementation. We read the FFLAGS, set the exception flags, and write the result back to the FFLAGS. This works in the gem5 simple and minor CPU model as they are actually written to `regFile` after executing the instructions. However, in the gem5 O3 CPU model, it will record in the `destMiscReg` buffer until the commit stage when writing to the `miscReg` in the execution stage. The next instruction will get the old FFLAGS and cause the incorrect result.

The CL introduced the `MISCREG_FFLAGS_EXE` and used the same size of `miscRegFile` because the `MISCREG_FFLAGS_EXE` and `MISCREG_FFLAGS` shared the same space. When executing the float-pointing instruction, any exception flags should be updated via `MISCREG_FFLAGS_EXE` to accumulate the FFLAGS in `setMiscReg` method. For the MISCREG_FFLAGS, it should only be called in the CSROp.

[1] Syntactic Dependencies: Appendix A
https://github.com/riscv/riscv-isa-manual/blob/c80ecada1ccc3120e2c8003bd7ca8e91c2e91bf9/src/mm-eplan.adoc#syntactic-dependencies-rules-9-11

Change-Id: Ib7f13d95b8a921c37766a54a217a5a4b1ef17c6f
Reviewed-on: https://soc-sim-external-review.googlesource.com/c/gem5/gem5/+/15141
Perf-Presubmit-Ready: Roger Chang <rogerycchang@google.com>
Gem5-Virtual-Platform-Presubmit-Ready: Roger Chang <rogerycchang@google.com>
Gem5-Virtual-Platform-Verified: kokoro <noreply+kokoro@google.com>
Reviewed-by: Roger Chang <rogerycchang@google.com>
Perf-Verified: kokoro <noreply+kokoro@google.com>